Decompose conenction classes
